requirements.txt
django-2.2.5
django-widget-tweaks-1.4.5
autopep8-1.4.4
